requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2025.11.12
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
filelock==3.20.1
frozenlist==1.8.0
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.15
idna==3.11
instructor==1.13.0
Jinja2==3.1.6
jiter==0.11.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
nodeenv==1.10.0
openai==2.14.0
platformdirs==4.5.1
pre_commit==4.5.1
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.8.1
tenacity==9.1.2
tqdm==4.67.1
ty==0.0.5
typer==0.20.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
virtualenv==20.35.4
yarl==1.22.0
